+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - AG-10, Add ability to create a new character
 - AG-24, Add sprite to character entity
 - AG-8, Connect to a new user configuration
+- AG-12, Add position to user model